Move ModelGitOps code into its own model_gitops.go
Follow-up for #973 and #989
